requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests
python-telegram-bot==13.10.0
SQLAlchemy==1.4.29
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans
google-trans-new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon==1.24.0
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==1.4.16
PyDictionary
yt-dlp
youtube_search_python
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
hiredis
pretty_errors
requests_html
ShazamAPI
html-telegraph-poster
qrcode
flask